+++ src/js/core/inverse.js
@@ -1,4 +1,14 @@
-import { $$, css, dimensions, matches, observeResize, on, replaceClass, toNodes } from 'uikit-util';
+import {
+    $$,
+    css,
+    dimensions,
+    matches,
+    observeResize,
+    on,
+    parent,
+    replaceClass,
+    toNodes,
+} from 'uikit-util';
 import { intersection, mutation } from '../api/observables';
 
 export default {
@@ -87,17 +97,19 @@ export default {
 
 function findTargetColor(target) {
     const { left, top, height, width } = dimensions(target);
+    const viewport = dimensions(window);
 
     let last;
     for (const percent of [0.25, 0.5, 0.75]) {
         const elements = target.ownerDocument.elementsFromPoint(
-            Math.max(0, left) + width * percent,
-            Math.max(0, top) + height / 2,
+            Math.max(0, Math.min(left + width * percent, viewport.width - 1)),
+            Math.max(0, Math.min(top + height / 2, viewport.height - 1)),
         );
 
         for (const element of elements) {
             if (
                 target.contains(element) ||
+                !checkVisibility(element) ||
                 (element.closest('[class*="-leave"]') &&
                     elements.some((el) => element !== el && matches(el, '[class*="-enter"]')))
             ) {
@@ -117,4 +129,20 @@ function findTargetColor(target) {
     }
 
     return last ? `uk-${last}` : '';
+}
+
+// TODO: once it becomes Baseline `element.checkVisibility({ opacityProperty: true, visibilityProperty: true })`
+function checkVisibility(element) {
+    if (css(element, 'visibility') !== 'visible') {
+        return false;
+    }
+
+    while (element) {
+        if (css(element, 'opacity') === '0') {
+            return false;
+        }
+        element = parent(element);
+    }
+
+    return true;
 }