+++ src/js/core/sticky.js
@@ -16,7 +16,6 @@ import {
     isVisible,
     noop,
     offsetPosition,
-    once,
     parent,
     query,
     remove,
@@ -97,32 +96,28 @@ export default {
     },
 
     observe: [
-        viewport({
-            handler() {
-                if (toPx('100vh', 'height') !== this._data.viewport) {
-                    this.$emit('resize');
-                }
-            },
-        }),
+        viewport(),
         scroll({ target: () => document.scrollingElement }),
         resize({
-            target: () => document.scrollingElement,
-            options: { box: 'content-box' },
+            target: ({ $el }) => [$el, parent($el), document.scrollingElement],
+            handler(entries) {
+                this.$emit(
+                    this._data.resized && entries.some(({ target }) => target === parent(this.$el))
+                        ? 'update'
+                        : 'resize',
+                );
+                this._data.resized = true;
+            },
         }),
-        resize(),
     ],
 
     events: [
         {
             name: 'load hashchange popstate',
 
-            el() {
-                return window;
-            },
+            el: () => window,
 
-            filter() {
-                return this.targetOffset !== false;
-            },
+            filter: ({ targetOffset }) => targetOffset !== false,
 
             handler() {
                 const { scrollingElement } = document;
@@ -146,29 +141,18 @@ export default {
                 });
             },
         },
-        {
-            name: 'transitionstart',
-
-            handler() {
-                this.transitionInProgress = once(
-                    this.$el,
-                    'transitionend transitioncancel',
-                    () => (this.transitionInProgress = null),
-                );
-            },
-        },
     ],
 
     update: [
         {
-            read({ height, width, margin, sticky }) {
+            read({ height, width, margin, sticky }, types) {
                 this.inactive = !this.matchMedia || !isVisible(this.$el);
 
                 if (this.inactive) {
                     return;
                 }
 
-                const hide = this.isFixed && !this.transitionInProgress;
+                const hide = this.isFixed && types.has('update');
                 if (hide) {
                     preventTransition(this.target);
                     this.hide();
@@ -243,6 +227,7 @@ export default {
                     top: offsetPosition(referenceElement)[0],
                     sticky,
                     viewport,
+                    maxScrollHeight,
                 };
             },
 
@@ -287,8 +272,9 @@ export default {
                 elHeight,
                 height,
                 sticky,
+                maxScrollHeight,
             }) {
-                const scroll = document.scrollingElement.scrollTop;
+                const scroll = Math.min(document.scrollingElement.scrollTop, maxScrollHeight);
                 const dir = prevScroll <= scroll ? 'down' : 'up';
                 const referenceElement = this.isFixed ? this.placeholder : this.$el;
 
@@ -490,7 +476,10 @@ function reset(el) {
     css(el, { position: '', top: '', marginTop: '', width: '' });
 }
 
-function preventTransition(el) {
-    addClass(el, 'uk-transition-disable');
-    requestAnimationFrame(() => removeClass(el, 'uk-transition-disable'));
+const clsTransitionDisable = 'uk-transition-disable';
+function preventTransition(element) {
+    if (!hasClass(element, clsTransitionDisable)) {
+        addClass(element, clsTransitionDisable);
+        requestAnimationFrame(() => removeClass(element, clsTransitionDisable));
+    }
 }